# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLoginDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "Created" },
                        "statusCode": { "type": "number", "example": 201 }
                      }
                    }
                  ],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthResponseDto" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.112Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreateDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "Created" },
                        "statusCode": { "type": "number", "example": 201 }
                      }
                    }
                  ],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthResponseDto" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.112Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Conflict" },
                        "statusCode": { "type": "number", "example": 409 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.113Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/refresh-token": {
      "get": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthRefreshResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Unauthorized" },
                        "statusCode": { "type": "number", "example": 401 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.113Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["auth"],
        "security": [{ "refreshToken": [] }]
      }
    },
...
    "/api/v1/notifications": {
      "post": {
        "operationId": "NotificationController_createNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NotificationCreateDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "Created" },
                        "statusCode": { "type": "number", "example": 201 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationCreateResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Conflict" },
                        "statusCode": { "type": "number", "example": 409 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.177Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["notifications"],
        "security": [{ "accessToken": [] }]
      },
      "get": {
        "operationId": "NotificationController_getNotifications",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "limit",
            "example": 10,
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page",
            "example": 0,
            "schema": { "type": "number" }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "description": "search",
            "example": "iste",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationGetResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.177Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["notifications"],
        "security": [{ "accessToken": [] }]
      }
    },
    "/api/v1/notifications/{id}": {
      "delete": {
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.177Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["notifications"],
        "security": [{ "accessToken": [] }]
      }
    }
  },
  "info": {
    "title": "nestjs-starter APIs Specification",
    "description": "Section for description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 201 },
          "message": { "type": "string", "nullable": false },
          "timestamp": {
            "type": "string",
            "example": "2024-04-18T08:48:35.068Z",
            "nullable": false
          },
          "data": { "type": "object", "nullable": false }
        },
        "required": ["statusCode", "message", "timestamp", "data"]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "avatar_id": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "is_verified": { "type": "boolean" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "object" },
          "updated_at": { "format": "date-time", "type": "string" },
          "password": { "type": "string" }
        },
        "required": [
          "avatar_id",
          "created_at",
          "deleted_at",
          "email",
          "first_name",
          "id",
          "is_deleted",
          "is_verified",
          "last_name",
          "phone",
          "role",
          "updated_at",
          "password"
        ]
      },
...
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "record is deleted" }
        },
        "required": ["status", "message"]
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "avatar_id": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "is_verified": { "type": "boolean" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "object" },
          "updated_at": { "format": "date-time", "type": "string" },
          "password": { "type": "string" }
        },
        "required": [
          "avatar_id",
          "created_at",
          "deleted_at",
          "email",
          "first_name",
          "id",
          "is_deleted",
          "is_verified",
          "last_name",
          "phone",
          "role",
          "updated_at",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "xE9iQWcOqV94jkOp8i3CH7zlwpklrM",
            "nullable": false
          },
          "refreshToken": {
            "type": "string",
            "example": "4vIHhd7IRgQErS4JvRsOaETsIwppXV",
            "nullable": false
          },
          "user": {
            "nullable": false,
            "allOf": [{ "$ref": "#/components/schemas/UserResponseDto" }]
          }
        },
        "required": ["accessToken", "refreshToken", "user"]
      },
      "UserLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "user email address",
            "example": "Camren26@yahoo.com"
          },
          "password": {
            "type": "string",
            "description": "user password",
            "example": "23tBw4TqfxfRB5K"
          }
        },
        "required": ["email", "password"]
      },
      "UserCreateDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "user email address",
            "example": "Camren26@yahoo.com"
          },
          "password": {
            "type": "string",
            "description": "user password",
            "example": "23tBw4TqfxfRB5K"
          },
          "firstName": {
            "type": "string",
            "description": "user first name",
            "example": "Samanta"
          },
          "lastName": {
            "type": "string",
            "description": "user last name",
            "example": "Stehr"
          }
        },
        "required": ["email", "password"]
      },
      "AuthRefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eThHQGH9gesqQYgRxCP0ZRaKVwPcx0",
            "nullable": false
          },
          "refreshToken": {
            "type": "string",
            "example": "1LL3r0EBlngrfT1M1YMuVWJ92BELy4",
            "nullable": false
          }
        },
        "required": ["accessToken", "refreshToken"]
      },
...
      "RecipientsDto": {
        "type": "object",
        "properties": {
          "created_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "user": { "$ref": "#/components/schemas/UserDto" },
          "notification_id": { "type": "string" },
          "users_id": { "type": "string" }
        },
        "required": [
          "created_at",
          "id",
          "updated_at",
          "user",
          "notification_id",
          "users_id"
        ]
      },
      "NotificationCreateResponseDto": {
        "type": "object",
        "properties": {
          "body": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "payload": { "type": "object" },
          "sender_id": { "type": "string" },
          "title": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "type": { "type": "object" },
          "sender": { "$ref": "#/components/schemas/UserDto" },
          "recipients": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RecipientsDto" }
          }
        },
        "required": [
          "body",
          "created_at",
          "deleted_at",
          "id",
          "is_deleted",
          "payload",
          "sender_id",
          "title",
          "updated_at",
          "type",
          "sender",
          "recipients"
        ]
      },
      "NotificationCreateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "title content",
            "example": "Deludo utique colligo amiculum depono modi coniuratio voluptas basium.\nCotidie altus consuasor fuga thermae."
          },
          "body": {
            "type": "string",
            "description": "notification body",
            "example": "repellat culpo aduro"
          },
          "payload": {
            "type": "object",
            "description": "payload",
            "example": {}
          },
          "recipients": {
            "description": "recipients",
            "example": [],
            "type": "array",
            "items": { "type": "string" }
          },
          "type": {
            "type": "object",
            "description": "notification type",
            "example": { "EMAIL": "EMAIL", "TEXT": "TEXT", "IN_APP": "IN_APP" }
          }
        },
        "required": ["title", "body", "type"]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "body": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "payload": { "type": "object" },
          "sender_id": { "type": "string" },
          "title": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "type": { "type": "object" },
          "sender": { "$ref": "#/components/schemas/UserDto" },
          "recipients": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RecipientsDto" }
          }
        },
        "required": [
          "body",
          "created_at",
          "deleted_at",
          "id",
          "is_deleted",
          "payload",
          "sender_id",
          "title",
          "updated_at",
          "type",
          "sender",
          "recipients"
        ]
      },
      "NotificationGetResponseDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "example": 10, "nullable": false },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NotificationDto" }
          }
        },
        "required": ["count", "data"]
      }
    }
  }
}